refactor(bootstrap): align wiring surface names and emitter ownership
Rename WiredDependencies.space_setup to sync_engine so the field matches
its SyncEngineDeps type and no longer collides nominally with the
unrelated AppFacade.space_setup (SpaceSetupFacade); update the
construction and access sites in wire.rs, non_gui.rs, and the daemon
bootstrap.

Relocate LoggingHostEventEmitter from entrypoint::non_gui to
observability::host_event so the shared wiring root no longer depends on
a specific scenario entrypoint (layering inversion).

Correct the SharedRuntimeDeps.host_event_bus doc: the bus is created
with the "logging" emitter already registered, not initially empty.
